+++ src/mappers/statsd.js
@@ -94,6 +94,10 @@ function mapEvents (prefix, data) {
 }
 
 function mapMetric (prefix, name, value) {
+    if (value <= 0) {
+        return '';
+    }
+
     return prefix + name + ':' + value + '|ms' + '\n';
 }